updated libft after submission

+++ ft_memcmp.c
@@ -1,18 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcmp.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kmurugan <kmurugan@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/10/16 18:23:33 by kmurugan          #+#    #+#             */
+/*   Updated: 2025/10/17 22:22:59 by kmurugan         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	const unsigned char *str1;
-	const unsigned char *str2;
+	const unsigned char	*str1;
+	const unsigned char	*str2;
 
 	str1 = (const unsigned char *)s1;
 	str2 = (const unsigned char *)s2;
-	while(n-- > 0)
+	while (n-- > 0)
 	{
 		if (*str1 != *str2)
-			return (*str1 - *str2);
+			return ((int)(*str1 - *str2));
 		str1++;
-		str2++;	
+		str2++;
 	}
-	return(0);
+	return (0);
 }
